Dhara library working
I have the dhara library working but will now create useful wrapper functions to make life easier when using the library. Also will attempt to make unit test functions for the library. Another thing that needs to be fixed is the usedSparesCount that is currently only accessible in the nand.c file and cant be changed from the dharaInit function.