Wrapper Work
moved the my_nand declaration to dharas nand.h file just to hide it from people accessing it. only dhara should uses the struct and thing else can use the defines located in the Dhara_Wrapper.h file. Currently just figuring out the wrap functions, will have to decide what to do when users demand for more data then a page and when more then a page is given to the write function.